Nano and LCD implementation
Swapped the main microprocessor board for a Arduino Nano instead of an ESP8266, as the ESP didn't have enough usable digital pins. Added the LCD display and tested it. Mounted various components inside the housing. Now just have to wire in, test and mount the BMP280, DS3231 and maybe a microSD card reader if I feel like it