Assembly and further testing
Final assembly complete. Considering adding the BMP280 pressure, temperature and humidity sensor and a microSD card reader in the future, but for now these are not present. The DS3231 has 32k of onboard addressable EEPROM and a temperature sensor, so these might negate the need for those other 2 modules.
Tested the DS3231 in this update - seems to work fine. Using [this](https://github.com/NorthernWidget/DS3231) library for the DS3231 as of now.
Decided not to include the BMP280 as it uses 3.3v logic on the I2C bus whereas the DS3231 and LCD use 5v.