Idle LCD code + MIDI
Idle code for the LCD added to the mix. The new sketch now allows the Arduino to spin every minute and update the display with the date, time and temperature accordingly.
Did lots of MIDI testing as well, tried to replicate Giorgio Moroder's "The Duel" from the 1984 classic "Electric Dreams". Works on but conks out at high notes and my MIDI conversion is far from perfect. Will definitely need to add microSD storage in the future.