MicroSD testing
Added a microSD card reader. Tested that it can at least read from the card. Followed a tutorial to play a .WAV file from the uSD. Works, even though it's far too quiet. Also added a second buzzer because I already had the wiring completed for it. It works too